moved auth API to service layer
